Frontend UI of dashboard, home page and food list as well as validation on customer, type and categories

+++ aspnet-core/src/OrderingSystem.Application/Categories/CategoriesAppService.cs
@@ -1,17 +1,20 @@
 ﻿using Abp.Application.Services;
 using Abp.Application.Services.Dto;
 using Abp.Domain.Repositories;
+using Abp.UI;
 using OrderingSystem.Categories.Dto;
 using OrderingSystem.Entities;
+using System.Linq;
 using System.Threading.Tasks;
 
 namespace OrderingSystem.Categories
 {
     public class CategoriesAppService : AsyncCrudAppService<Category, CategoriesDto, int, PagedCategoriesResultRequestDto, CreateCategoriesDto, CategoriesDto>, ICategoriesAppService
     {
-        public CategoriesAppService(IRepository<Category, int> repository) : base(repository)
+        private readonly IRepository<Food, int> _foodrepository;
+        public CategoriesAppService(IRepository<Category, int> repository, IRepository<Food, int> foodrepository) : base(repository)
         {
-
+            _foodrepository = foodrepository;
         }
 
         public override Task<CategoriesDto> CreateAsync(CreateCategoriesDto input)
@@ -21,7 +24,17 @@ namespace OrderingSystem.Categories
 
         public override Task DeleteAsync(EntityDto<int> input)
         {
-            return base.DeleteAsync(input);
+            var exist = _foodrepository.GetAll()
+                .Where(x => x.CategoryId == input.Id)
+                .FirstOrDefault();
+            if (exist == null)
+            {
+                return base.DeleteAsync(input);
+            }
+            else
+            {
+                throw new UserFriendlyException("Category exist in the list of foods");
+            }
         }
 
         public override Task<PagedResultDto<CategoriesDto>> GetAllAsync(PagedCategoriesResultRequestDto input)
